feat: Implement license check and increase QR code limit
This commit introduces a license check to the home page and increases the maximum allowed QR codes from 3500 to 5000.

The license check is implemented using a new `LicenseService` to retrieve and display license information on the welcome page.

The QR code limit is increased to accommodate larger datasets. Also, the pdf name is changed to remove the 'bizli_labels_' prefix.